requirements.txt
numpy==1.12.1
protobuf==3.3.0
scipy==0.19.0
six==1.10.0
tensorflow-gpu==1.1.0
Werkzeug==0.12.2
